License changes
This library may be used by anyone in any environment. If someone finds
it useful to change someting, he should give back to the community by
making the changes available for everyone. This is why the GPL is used
as license. If someday there is a commercial support service, there
might be reasons to allow commercial licenses.