Basic initialization
refs NET-4

It seems to be clear that a socket of a newer type has to have support
for IPv6 and let the OS handle the mapping to IPv4 if necessary. Hence
the Socket is initialized as UDPv6 socket and is bound to a user
specified port for any address that messages com from.

The socket is also told to be not blocking, though this might be
unnecessary since this should work in an asynchronous way.