Switching to managed socket
When deriving from the socket class, implicitly all its public methods
are exposed. But we want to have control over the API, thus we
encapsulate the socket and offer only relevant information.

The peer socket now finds out about the MTU of the NICs on the host.
This will determine the buffersize for the SAEA objects, since we can
not receive messages that have a bigger size then that.

Exposing exceptions is avoided by catching them all in the required
places.

NET-4, NET-9