Unreferencing event listeners
I read an article today that stated that event listeners never get freed
if you don't clean up after your class. So you will be probably carrying
a lot of garbage around if they are are not freed. And since this
library wants to make everyones life easier, we take care of this for
you.

The article I read was:

http://www.codeproject.com/Articles/639493/Events-Demystifying-Common-Memory-Leaks