Separating buffers
I introduced a scond queue of buffers dedicated to sending messages.
This doubles the buffersize allocated but makes sure that sending can't
fail due to no more SAEA objects in the _stateobjects pool.

Also the SAEA objects are now unregistering the complete methods.

The buffer size now is not only determined by the MTU but takes IP- and
UDP- headers into account.

NET-4