Introducing the Message Router
This class is going to be the first one that actually looks at the
contents of an incoming message. An application that is trying to work
with this library will have to register a protocol ID that it is
expecting. All messages that have this protocol ID will get forwarded to
the application and can be parsed further. The protocol ID is stripped
from the packet that is handed over to the application.

But this is just a stub yet ;)