Application interface and signaling added
I have created an Interface that all applications using the library havt
ot implement. It is holding the protocolID of the application and a
method that is called with incoming data. I also tried to trigger events
from the message router on an application, but it is failing due to the
false assignement of the integer representing the protocolID on the
receiver side.